+++ bitrix/modules/sale/admin/delivery_eservice_list.php
@@ -27,31 +27,32 @@ namespace Bitrix\Sale\Delivery\AdminPage\DeliveryExtraServiceEdit
 	Loc::loadMessages(__FILE__);
 
 	$ID = intval($_GET['ID']);
-	$strError = "";
+	global $srvStrError;
 
-	if(isset($_REQUEST["action"]) && $_REQUEST["action"] == "delete_extra_service" && isset($_REQUEST["ES_ID"]))
+	if(isset($_REQUEST["action"]) && $_REQUEST["action"] == "delete_extra_service" && isset($_REQUEST["ES_ID"]) && $saleModulePermissions == "W" && check_bitrix_sessid())
 	{
-		if(intval($_REQUEST["ES_ID"] > 0))
+		if(intval($_REQUEST["ES_ID"]) > 0)
 		{
 			$res = ExtraServices\Table::delete(intval($_REQUEST["ES_ID"]));
 
 			if(!$res->isSuccess())
-				$strError .= Loc::getMessage("SALE_ESDE_ERROR_DELETE").' '.implode("<br>\n",$res->getErrorMessages());
+				$srvStrError .= Loc::getMessage("SALE_ESDE_ERROR_DELETE").' '.implode("<br>\n",$res->getErrorMessages());
 		}
 		else
 		{
-			$strError .= Loc::getMessage("SALE_ESDE_ERROR_ID");
+			$srvStrError .= Loc::getMessage("SALE_ESDE_ERROR_ID");
 		}
 	}
 
 	$tableId = 'table_delivery_extra_service';
 	$oSort = new \CAdminSorting($tableId);
 	$lAdmin = new \CAdminList($tableId, $oSort);
+	$esClasses = ExtraServices\Manager::getClassesList();
 
 	$res = \Bitrix\Sale\Delivery\ExtraServices\Table::getList(array(
 		'filter' => array(
-			'DELIVERY_ID' => $ID,
-			'=CLASS_NAME' => ExtraServices\Manager::getClassesList()
+			'=DELIVERY_ID' => $ID,
+			'=CLASS_NAME' => $esClasses
 		),
 		'select' => array('ID', 'CODE', 'NAME', 'DESCRIPTION', 'CLASS_NAME', 'RIGHTS', 'ACTIVE', 'SORT'),
 		'order' => array('SORT' => 'ASC', 'ID' => 'DESC')
@@ -82,9 +83,9 @@ namespace Bitrix\Sale\Delivery\AdminPage\DeliveryExtraServiceEdit
 		$link = 'sale_delivery_eservice_edit.php?ID='.$record['ID'].'&lang='.LANGUAGE_ID.'&'.$tabControl->ActiveTabParam().'&back_url='.urlencode($APPLICATION->GetCurPageParam());
 		$row =& $lAdmin->AddRow($record['ID'], $record, $link, '');
 		$row->AddField('ID', '<a href="'.$link.'">'.$record['ID'].'</a>');
-		$row->AddField('CODE', $record['CODE']);
-		$row->AddField('NAME', $record['NAME']);
-		$row->AddField('SORT', $record['SORT']);
+		$row->AddField('CODE', htmlspecialcharsbx($record['CODE']));
+		$row->AddField('NAME', htmlspecialcharsbx($record['NAME']));
+		$row->AddField('SORT', intval($record['SORT']));
 		$row->AddField('RIGHTS', $record['RIGHTS']);
 		$row->AddField('ACTIVE', $record['ACTIVE'] == "Y" ? Loc::getMessage('SALE_ESDL_YES') : Loc::getMessage('SALE_ESDL_NO'));
 
@@ -94,7 +95,7 @@ namespace Bitrix\Sale\Delivery\AdminPage\DeliveryExtraServiceEdit
 			$className = "";
 
 		$row->AddField('CLASS_NAME', $className);
-		$row->AddField('DESCRIPTION', $record['DESCRIPTION']);
+		$row->AddField('DESCRIPTION', htmlspecialcharsbx($record['DESCRIPTION']));
 
 		if ($saleModulePermissions >= "W")
 		{
@@ -131,9 +132,10 @@ namespace Bitrix\Sale\Delivery\AdminPage\DeliveryExtraServiceEdit
 			"ICON" => "btn_new"
 		);
 
+		$menu = array();
+
 		if($service && $embeddedList = $service->getEmbeddedExtraServicesList())
 		{
-			$menu = array();
 
 			foreach($embeddedList as $code => $eserviceParams)
 			{
@@ -143,17 +145,29 @@ namespace Bitrix\Sale\Delivery\AdminPage\DeliveryExtraServiceEdit
 				);
 			}
 
-			if(!empty($menu))
-			{
-				array_unshift($menu, array(
-					'TEXT' => Loc::getMessage('SALE_ESDL_NEW_SERVICE'),
-					"LINK" => 'sale_delivery_eservice_edit.php?lang='.LANGUAGE_ID.'&DELIVERY_ID='.$ID.'&'.$tabControl->ActiveTabParam().'&back_url='.urlencode($APPLICATION->GetCurPageParam()),
-				));
+			sortByColumn($menu, array("TEXT" => SORT_ASC));
 
-				$addButtonParams["MENU"] = $menu;
-			}
+			$menu[] =  array(
+				'SEPARATOR' => true,
+			);
 		}
 
+		/** @var  \Bitrix\Sale\Delivery\ExtraServices\Base $esClass */
+		foreach(ExtraServices\Manager::getClassesList() as $esClass)
+		{
+			if($esClass == '\Bitrix\Sale\Delivery\ExtraServices\String')
+				continue;
+
+			if($esClass::isEmbeddedOnly())
+				continue;
+
+			$menu[] =  array(
+				'TEXT' => $esClass::getClassTitle(),
+				"LINK" => 'sale_delivery_eservice_edit.php?lang='.LANGUAGE_ID.'&DELIVERY_ID='.$ID.'&'.$tabControl->ActiveTabParam().'&CLASS_NAME='.urlencode($esClass).'&back_url='.urlencode($APPLICATION->GetCurPageParam()),
+			);
+		}
+
+		$addButtonParams["MENU"] = $menu;
 		$aContext[] = $addButtonParams;
 		$lAdmin->AddAdminContextMenu($aContext, false);
 	}
